refactor(profile): simplify the creation code
